# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Course memo/Kurs-PM API",
    "description": "Api to store information about uploaded course memos to be used in course development and course info projects",
    "version": "1.0.0"
  },
  "basePath": "/api/kurs-pm",
  "produces": ["application/json"],
  "paths": {
...
    "/v1/usedRoundsForCourse/{courseCode}/semester/{semester}": {
      "get": {
        "operationId": "getUsedRounds",
        "summary": "Gets a list of used round ids for a semester in a course",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "Course code, six characters",
            "required": true,
            "type": "string"
          },
          {
            "name": "semester",
            "description": "Semester for rounds",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "Round ids",
            "schema": {
              "$ref": "#/definitions/UsedRoundsForCourseAndSemester"
            }
          },
          "401": {
            "description": "This endpoint is secured - **credentials are missing or invalid**."
          },
          "400": {
            "description": "Bad Request, something with the request is invalid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No organization or user matched the Ids",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header",
      "scopes": {
        "read": "Read access to data",
        "write": "Write access to data"
      }
    }
  },
  "definitions": {
...
    "UsedRoundsForCourseAndSemester": {
      "roundIdList": {
        "type": "string",
        "description": "List course rounds"
      },
      "memoList": {
        "type": "string",
        "description": "Who made changes to documents"
      }
    },
    "Error": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    },
...
  }
}